Refactor outstanding queue helpers
