+++ src/ui/dashboard/quickActions.js
@@ -1,6 +1,6 @@
 import { formatHours } from '../../core/helpers.js';
 import { clampNumber } from './formatters.js';
-import { collectOutstandingActionEntries } from '../actions/registry.js';
+import { collectOutstandingActionEntries } from '../actions/outstanding.js';
 import { buildQueueMetrics } from '../actions/queueService.js';
 import { registerActionProvider } from '../actions/providers.js';
 import { executeAction } from '../../game/actions.js';
