+++ tests/ui/update.integration.test.js
@@ -862,8 +862,8 @@ test('accepting a hustle offer from dashboard quick actions refreshes the todo q
   const instanceId = acceptedEntries[0]?.instanceId;
   assert.ok(instanceId, 'expected claimed hustle entry to include an instance id');
 
-  const actionsRegistry = await import('../../src/ui/actions/registry.js');
-  const outstandingEntries = actionsRegistry.collectOutstandingActionEntries(state);
+  const outstandingModule = await import('../../src/ui/actions/outstanding.js');
+  const outstandingEntries = outstandingModule.collectOutstandingActionEntries(state);
   const hasOutstanding = outstandingEntries.some(entry => entry?.id === `instance:${instanceId}`);
   assert.ok(hasOutstanding, 'expected outstanding entry for the accepted hustle');
